# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "仿抖音实现视频的浏览，评论，喜欢，发布等功能",
        "title": "tiktok抖音项目",
        "contact": {
            "name": "konyue,msa52412,Mrxuexi, 199094212",
            "url": "https://github.com/Mrxuexi/tiktok"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/Mrxuexi/tiktok/blob/main/LICENSE.txt"
        },
        "version": "0.0.2"
    },
    "host": "127.0.0.1:8080",
    "paths": {
        "/douyin/comment/action": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "返回操作是否成功",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "评论"
                ],
                "summary": "评论请求",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "actionType",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "commentId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "commentText",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "videoId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/io.CommentActionResponse"
                        }
                    }
                }
            }
        },
        "/douyin/comment/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "返回评论列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "评论"
                ],
                "summary": "获取评论列表",
                "parameters": [
                    {
                        "type": "string",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "videoId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/io.CommentListResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "io.Comment": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "create_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/io.User"
                }
            }
        },
        "io.CommentActionResponse": {
            "type": "object",
            "properties": {
                "comment": {
                    "$ref": "#/definitions/io.Comment"
                },
                "status_code": {
                    "type": "integer"
                },
                "status_msg": {
                    "type": "string"
                }
            }
        },
        "io.CommentListResponse": {
            "type": "object",
            "properties": {
                "comment_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/io.Comment"
                    }
                },
                "status_code": {
                    "type": "integer"
                },
                "status_msg": {
                    "type": "string"
                }
            }
        },
        "io.User": {
            "type": "object",
            "properties": {
                "follow_count": {
                    "description": "关注总数",
                    "type": "integer"
                },
                "follower_count": {
                    "description": "粉丝总数",
                    "type": "integer"
                },
                "id": {
                    "description": "用户id",
                    "type": "integer"
                },
                "is_follow": {
                    "description": "true-已关注，false-未关注",
                    "type": "boolean"
                },
                "name": {
                    "description": "用户名称",
                    "type": "string"
                }
            }
        }
    }
}